requirements.txt
pymupdf
sentence-transformers 
faiss-cpu